refactor databricks auth
